(PC-17816) refactor(profile): move useValidateEmail and useCheckHasCurrentEmailChange from pages to helpers
