+++ src/features/profile/pages/ChangeEmail/ChangeEmail.tsx
@@ -12,10 +12,10 @@ import { isLongEnough } from 'features/auth/components/PasswordSecurityRules'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { ChangeEmailRequest, CHANGE_EMAIL_ERROR_CODE } from 'features/profile/api'
+import { useCheckHasCurrentEmailChange } from 'features/profile/helpers/useCheckHasCurrentEmailChange'
+import { useValidateEmail } from 'features/profile/helpers/useValidateEmail'
 import { AlreadyChangedEmailDisclaimer } from 'features/profile/pages/ChangeEmail/AlreadyChangedEmailDisclaimer'
 import { ChangeEmailDisclaimer } from 'features/profile/pages/ChangeEmail/ChangeEmailDisclaimer'
-import { useCheckHasCurrentEmailChange } from 'features/profile/pages/ChangeEmail/utils/useCheckHasCurrentEmailChange'
-import { useValidateEmail } from 'features/profile/pages/ChangeEmail/utils/useValidateEmail'
 import { analytics } from 'libs/firebase/analytics'
 import { useSafeState } from 'libs/hooks'
 import { theme } from 'theme'
